requirements.txt
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.6.1
PyYAML==3.11
RPi.GPIO==0.6.2
RTIMULib==7.2.1
Werkzeug==0.11.11
argparse==1.2.1
chardet==2.3.0
click==6.6
colorama==0.3.2
gpiozero==1.2.0
html5lib==0.999
imutils==0.3.6
itsdangerous==0.24
lxkeymap==0.1
mcpi==0.1.1
ndg-httpsclient==0.3.2
numpy==1.8.2
picamera==1.10
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.30
pyOpenSSL==0.13.1
pyasn1==0.1.7
pygame==1.9.2a0
pygobject==3.14.0
pyserial==2.6
python-apt==0.9.3.12
requests==2.4.3
sense-hat==2.1.0
six==1.8.0
spidev==3.0
urllib3==1.9.1
virtualenv==15.0.2
wheel==0.24.0
wsgiref==0.1.2
